Rename JSON string array helper CIRC-381
